requirements.txt

scikit-learn
nltk
langchain>=0.1.13
langchain-community
langchain-groq==0.1.5
langchain-huggingface
faiss-cpu
sentence-transformers
python-dotenv
tqdm
streamlit
